+++ fiddles/javac/fiddle-0000-Template/template/src/Fiddle.java
@@ -1,0 +1,6 @@
+
+class Fiddle {
+    public static void main(String[] args) {
+        System.out.println("template compiled!");
+    }
+}